basic version using H2 database backend
So many weird little facts learned!

@Value variables are initialized AFTER the constructor, but there is a @PostConstructor annotation

gradle has a build cache that it tries to use, even when you say "gradle clean". That's pretty stupid and I think can be overridded in settings.gradle. Ran into problems with this.

Classes managed for persistence have to be in the same package, or a subackage. So I can't have:

caldera.Event

and then have server and client code in:

caldera.server.*
caldera.client.*

You can either move the class (which is what I did) or figure out how to set another annotation to tell Spring to look someplace else. Ugh.